Fix issue #131 LLFIO no longer compiles for Windows on clang 18.
Also fixed up lots of diagnostics if targeting Windows with clang.

+++ example/tls_socket_server.cpp
@@ -175,7 +175,7 @@ void tls_socket_server()
         }
         if(sock.wbuf.empty()) {
           // Nothing more to write, so no longer poll for this
-          topoll[handleidx]&=~llfio::poll_what::is_writable;
+          topoll[handleidx]&=uint8_t(~llfio::poll_what::is_writable);
         }
       }
